Website
Dear Instructors,
I totally forgot that we had to submit the project via Github and I accidentally submitted the project via the portal initially. I understand that it was completely my responsiblity to submit the code on Github, but I put a lot of work into this project so I was hoping you guys could still take a look at it and maybe dock off a few lateness points for my mistake. Thank you for your time and consideration.